+++ app/admin/admin_user.rb
@@ -1,0 +1,24 @@
+ActiveAdmin.register AdminUser do
+
+  menu priority: 4, label: 'Administrateurs'
+
+  config.filters = false
+
+  actions :all, except: [:delete, :show]
+
+  index do
+    column :username
+    actions
+  end
+
+  form do |f|
+    f.inputs do
+      f.input :username
+      f.input :password
+      f.input :password_confirmation
+    end
+
+    f.actions
+  end
+
+end