+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- VMs date's year now ranges from Time.zone.today.year - 1 - 1
+
 1.3.12
 ------