+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - give more space for "backup reason"
 
+- VMs date's year now ranges from Time.zone.today.year - 1 - 1
+
 1.3.12
 ------