+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - give more space for "backup reason"
 - VMs date's year now ranges from Time.zone.today.year - 1 - 1
 
+- add 'Email' to MaintenanceTicket::MAINTAINED_BY
+
 1.3.12
 ------