+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 - add 'Email' to MaintenanceTicket::MAINTAINED_BY
 - HACK to fix the vm creation to update the disks attributes
 
+1.4.0
+------
+
+- add admin_user admin
+
 1.3.12
 ------